Add some standardizations before actual formatting, and some actual formatting work.
The standardizations are almost like passes in a compiler, where passes depend on previous passes.

The standardizations are: Remove Line Continuations, Standardize Newlines, and Standardize Tabs.

# Remove Line Continuations

All line continuations are removed. Line continuations may occur in any LeafNodes, in any location within a LeafNode.

The backslash, newline, and any whitespace immediately following the newline are all removed.

# Standardize Newlines

All newlines are standardized to a single newline. The strings "\r\n", "\n", and "\r" are replaced with a single newline.

Newlines may occur within strings or comments

# Standardize Tabs

Tabs are replaced with the correct number of spaces, based on tab stops. This is needed to correctly compute columns for future formatting work. This is unrelated to "IndentationString" option that can be used during formatting to recreate indenting with tabs.

In actual formatting:

Correctly format comments, taking into account the existing formatting of the comment.

Provide an indent argument to line[].