Lots of work.
Define separate values for defaults of options

Add 2 new "passes": IntroduceRowNodes and AbstractFormatNodes
IntroduceRowNodes:
RowNodes are used to enforce nodes being on the same line. Right now this is only done for comments.

AbstractFormatNodes:
Some nodes can be "abstracted" for formatting
e.g.,

f[] := f[] = x

is a SetDelayed node, with a Set node on the RHS

But this really should be formatted as if it were a single TernaryNode

Handle more cases with GroupNodes, handle e.g., empty group nodes.

Remove newlines in variables list for Module, With, Block, Function. These newlines are added for line breaking purposes and will be redone.